Move sources logic inside trait
